+++ resources/views/Components/footer.blade.php
@@ -4,7 +4,7 @@
     <div class="footerColumnTwo">
         <hr style="visibility:hidden;" />
         <img src="{{ asset('images/logo.png') }}" class="footerLogo">
-        <p class="footerTextWhite">2022-2023 &#169</p>
+        <p class="footerTextWhite">2022-2024 &#169</p>
     </div>
 
     <div class="footerColumnFour">
@@ -16,7 +16,7 @@
     <div class="footerColumnFive">
         <p class="footerTextGrey">{!! __('Помощь') !!}</p>
         <p class="footerTextWhite">
-            <span style="color:yellow">8983125706a@gmail.com</span>
+            <span style="color:yellow">a.evtifeev@almau.edu.kz</span>
             - {!! __('при возникновении проблем или вопросов касательно работы сайта') !!}
         </p>
     </div>