+++ resources/views/Components/navBar.blade.php
@@ -5,6 +5,7 @@
     <div class="nav-bar-center">
         <a href="/SupportTicketStatus">{!! __('Статус заявки') !!}</a>
         <a href="/SupportTicketNew" class="red">{!! __('Подать заявку') !!}</a>
+        <a href="/DormitoryTicketNew">{!! __('Запись в общежитие') !!}</a>
     </div>
     <div class="nav-bar-right">
         <img src="{{ asset('images/language.png') }}" class="topLanguage">